Move the NullFeedMetric too where it is used.
